refactor(core): mxbsecure belongs in core, like sidecar
Placing the private modules for real is what showed this: both binaries use
mxbsecure. The manager reads sealed archives and injects the client at launch;
the studio locks content and seals a track to a buyer. It was declared in the
manager alone, so the studio's `mxbsecure_generate` was compiling only because
its `cfg(mxbsecure)` arm was never taken.

So it sits beside `sidecar` in mxb-core, and core's build script publishes both
decisions — `cargo::metadata=mxbsecure=…` alongside the sidecar one — with each
app mirroring rather than looking for a file it no longer holds. The DLL still
goes to the manager: injection is its job.

This is the first time the full-feature build has compiled since the split, and
turning that code path on found two things it had been hiding: `sanitize_asset_id`
was manager-local but is needed by the studio's sealing command, so it moved to
`mxb_core::names`; and `mod mxbsecure` was then declared twice in the manager.

release.yml drops the stopgap that folded sync.sh's output into apps/manager —
sync.sh knows the monorepo layout now (Frostn1/mxbapp-private ac50ba6) and places
each module where it belongs. The DLL check still points at the manager, which is
still where the DLL goes.

Verified with the private modules present, which is what makes this meaningful:
cargo check clean on host and x86_64-pc-windows-gnu; 914 + 367 + 213 = 1494 tests
pass. The six above the usual 1488 are the ones that only exist when these files
are there — including `sidecar::tests::decrypts_kcol_fixture` and the mxbsecure
offline-flow test, both passing.
