+++ crates/core/.gitignore
@@ -1,3 +1,4 @@
 # Local-only module (not part of the public tree). Placed by the private repo's sync.sh.
 /src/sidecar.rs
 /src/fixtures/test_sidecar.pkz
+/src/mxbsecure.rs
