+++ crates/core/Cargo.toml
@@ -50,6 +50,7 @@ reqwest = { version = "0.12", default-features = false, features = [
 ] }
 tokio = { version = "1", features = ["time", "sync", "rt"] }
 uuid = { version = "1", features = ["v4"], optional = true }
+sha2 = "0.10"
 base64 = "0.22"
 log = "0.4"
 dirs-next = "2.0.0"
