+++ crates/core/build.rs
@@ -16,4 +16,15 @@ fn main() {
     // land as a compile error deep inside sidecar_lock with nothing pointing at the cause.
     // Delivered as DEP_MXBCORE_SIDECAR, which the `links` key in Cargo.toml enables.
     println!("cargo::metadata=sidecar={}", if have { "1" } else { "0" });
+
+    // The secure-content packer, gated independently. Both binaries need it — the manager
+    // injects the client at launch, the studio seals a track to a buyer — so like `sidecar`
+    // it lives here and the decision is published rather than made twice.
+    println!("cargo::rustc-check-cfg=cfg(mxbsecure)");
+    let secure = Path::new("src/mxbsecure.rs").exists();
+    if secure {
+        println!("cargo::rustc-cfg=mxbsecure");
+    }
+    println!("cargo::rerun-if-changed=src/mxbsecure.rs");
+    println!("cargo::metadata=mxbsecure={}", if secure { "1" } else { "0" });
 }
